Wip: unit tests

+++ src/modules/schedules/useCases/deleteSchedule/DeleteScheduleUseCase.ts
@@ -1,6 +1,7 @@
 import { inject, injectable } from "tsyringe";
 
 import { ISchedulesRepository } from "@modules/schedules/repositories/ISchedulesRepository";
+import { AppError } from "@shared/errors/AppError";
 
 @injectable()
 class DeleteScheduleUseCase {
@@ -10,6 +11,11 @@ class DeleteScheduleUseCase {
   ) {}
 
   public async execute(id: string): Promise<void> {
+    const schedule = await this.schedulesRepository.findById(id);
+
+    if (!schedule) {
+      throw new AppError({ error: "Schedule doesn't exists" }, 400);
+    }
     await this.schedulesRepository.delete(id);
   }
 }
